pyproject.toml
# Apache License
#
# Copyright (c) 2023 Advanced Micro Devices, Inc.
#
# Permission is hereby granted, free of charge, to any person obtaining a copy
# of this software and associated documentation files (the "Software"), to deal
# in the Software without restriction, including without limitation the rights
# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
# copies of the Software, and to permit persons to whom the Software is
# furnished to do so, subject to the following conditions:
#
# The above copyright notice and this permission notice shall be included in all
# copies or substantial portions of the Software.
#
# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
# SOFTWARE.

[build-system]
build-backend = "setuptools.build_meta"
requires = [
    "setuptools>=64.0.0",
    "tomli ; python_version < '3.11'",
] # This list was generated by `rapids-dependency-file-generator`. To make changes, edit dependencies.yaml and run `rapids-dependency-file-generator`.

[project]
name = "dask-cuda"
dynamic = ["version"]
description = "Utilities for Dask and CUDA interactions"
readme = { file = "README.md", content-type = "text/markdown" }
authors = [
    { name = "NVIDIA Corporation" },
    { name = "AMD" },
]
license = { text = "Apache-2.0" }
requires-python = ">=3.9"
dependencies = [
    "click >=8.1",
    "numba>=0.57",
    "pyamdsmi >= 0.2.0",
    "numpy>=1.21",
    "pandas>=1.3",
    "pynvml>=11.0.0,<11.5",
    "rapids-dask-dependency==24.4.*",
    "zict>=2.0.0",
] # This list was generated by `rapids-dependency-file-generator`. To make changes, edit dependencies.yaml and run `rapids-dependency-file-generator`.
classifiers = [
    "Intended Audience :: Developers",
    "Topic :: Database",
    "Topic :: Scientific/Engineering",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]

[project.scripts]
dask-cuda-worker = "dask_cuda.cli:worker"
dask-cuda-config = "dask_cuda.cli:config"

[project.entry-points.dask_cli]
cuda = "dask_cuda.cli:cuda"

[project.optional-dependencies]
docs = [
    "numpydoc>=1.1.0",
    "sphinx",
    "sphinx-click>=2.7.1",
    "sphinx-rtd-theme>=0.5.1",
] # This list was generated by `rapids-dependency-file-generator`. To make changes, edit dependencies.yaml and run `rapids-dependency-file-generator`.
test = [
    "cudf==24.4.*",
    "dask-cudf==24.4.*",
    "kvikio==24.4.*",
    "pytest",
    "pytest-cov",
    "ucx-py==0.37.*",
] # This list was generated by `rapids-dependency-file-generator`. To make changes, edit dependencies.yaml and run `rapids-dependency-file-generator`.

[project.urls]
Homepage = "https://github.com/rapidsai/dask-cuda"
Documentation = "https://docs.rapids.ai/api/dask-cuda/stable/"
Source = "https://github.com/rapidsai/dask-cuda"

[tool.coverage.run]
disable_warnings = [
    "include-ignored",
]
include = [
    "dask_cuda/*",
]
omit = [
    "dask_cuda/tests/*",
]

[tool.isort]
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
combine_as_imports = true
order_by_type = true
known_dask = [
    "dask",
    "distributed",
]
known_rapids = [
    "rmm",
    "cuml",
    "cugraph",
    "dask_cudf",
    "cudf",
    "ucp",
]
known_first_party = [
    "dask_cuda",
]
default_section = "THIRDPARTY"
sections = [
    "FUTURE",
    "STDLIB",
    "THIRDPARTY",
    "DASK",
    "RAPIDS",
    "FIRSTPARTY",
    "LOCALFOLDER",
]
skip = [
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".tox",
    ".venv",
    "build",
    "dist",
    "__init__.py",
]

[tool.pytest.ini_options]
filterwarnings = [
    "error::DeprecationWarning",
    "error::FutureWarning",
    # remove after https://github.com/rapidsai/dask-cuda/issues/1087 is closed
    "ignore:There is no current event loop:DeprecationWarning:tornado",
    # This warning must be filtered until dask-expr support
    # is enabled in both dask-cudf and dask-cuda.
    # See: https://github.com/rapidsai/dask-cuda/issues/1311
    "ignore:Dask DataFrame implementation is deprecated:DeprecationWarning",
]

[tool.setuptools]
license-files = ["LICENSE"]

[tool.setuptools.dynamic]
version = {file = "dask_cuda/VERSION"}

[tool.setuptools.packages.find]
exclude = [
    "docs",
    "tests",
    "docs.*",
    "tests.*",
]
